Tooltips and refactoring of UI tests
