+++ tests/functional/tags_ui/test_filter_tag_ui.py
@@ -12,16 +12,18 @@ from src.models.utub_tags import Utub_Tags
 from src.models.utub_url_tags import Utub_Url_Tags
 from src.models.utub_urls import Utub_Urls
 from src.utils.strings.ui_testing_strs import UI_TEST_STRINGS as UTS
-from tests.functional.locators import HomePageLocators as HPL
-from tests.functional.tags_ui.utils_for_test_tag_ui import (
+from tests.functional.db_utils import (
     add_tag_to_utub_user_created,
     add_two_tags_across_urls_in_utub,
-    apply_tag_filter_by_id_and_get_shown_urls,
-    get_utub_tag_filter_selector,
-)
-from tests.functional.utils_for_test import (
     get_utub_this_user_created,
-    login_user_and_select_utub_by_utubid,
+)
+from tests.functional.locators import HomePageLocators as HPL
+from tests.functional.login_utils import login_user_and_select_utub_by_utubid
+from tests.functional.tags_ui.selenium_utils import (
+    apply_tag_filter_by_id_and_get_shown_urls,
+    get_utub_tag_filter_selector
+)
+from tests.functional.selenium_utils import (
     wait_then_click_element,
     wait_until_visible_css_selector,
 )
@@ -29,7 +31,6 @@ from tests.functional.utils_for_test import (
 pytestmark = pytest.mark.tags_ui
 
 
-# Happy Path Tests
 def test_filter_tag_with_all_urls_filtered(
     browser: WebDriver, create_test_urls, provide_app: Flask
 ):
