+++ tests/functional/utubs_ui/test_mobile_utubs_ui.py
@@ -4,19 +4,21 @@ from selenium.webdriver.remote.webdriver import WebDriver
 from src.cli.mock_constants import MOCK_UTUB_DESCRIPTION
 from src.models.utubs import Utubs
 from src.utils.strings.ui_testing_strs import UI_TEST_STRINGS as UTS
+from tests.functional.assert_utils import assert_panel_visibility_mobile
+from tests.functional.db_utils import get_utub_this_user_created
 from tests.functional.locators import HomePageLocators as HPL
-from tests.functional.utils_for_test import (
-    Decks,
-    click_on_navbar,
-    get_utub_this_user_created,
+from tests.functional.login_utils import (
     login_user_and_select_utub_by_utubid_mobile,
     login_user_to_home_page,
+)
+from tests.functional.selenium_utils import (
+    Decks,
+    click_on_navbar,
     select_utub_by_id_mobile,
-    verify_panel_visibility_mobile,
     wait_then_click_element,
     wait_until_hidden,
 )
-from tests.functional.utubs_ui.utils_for_test_utub_ui import create_utub
+from tests.functional.utubs_ui.selenium_utils import create_utub
 
 pytestmark = pytest.mark.mobile_ui
 
@@ -46,7 +48,7 @@ def test_create_utub_shows_url_deck_mobile(
     # Wait for POST request
     wait_until_hidden(browser, HPL.INPUT_UTUB_NAME_CREATE, timeout=3)
 
-    verify_panel_visibility_mobile(browser=browser, visible_deck=Decks.URLS)
+    assert_panel_visibility_mobile(browser=browser, visible_deck=Decks.URLS)
 
 
 def test_tap_on_already_selected_utub_shows_url_deck_mobile(
@@ -66,13 +68,13 @@ def test_tap_on_already_selected_utub_shows_url_deck_mobile(
     login_user_and_select_utub_by_utubid_mobile(
         app=app, browser=browser, utub_id=utub.id, user_id=USER_ID
     )
-    verify_panel_visibility_mobile(browser=browser, visible_deck=Decks.URLS)
+    assert_panel_visibility_mobile(browser=browser, visible_deck=Decks.URLS)
 
     # Travel back
     click_on_navbar(browser)
 
     wait_then_click_element(browser, HPL.NAVBAR_UTUB_DECK, time=10)
-    verify_panel_visibility_mobile(browser=browser, visible_deck=Decks.UTUBS)
+    assert_panel_visibility_mobile(browser=browser, visible_deck=Decks.UTUBS)
 
     select_utub_by_id_mobile(browser, utub.id)
-    verify_panel_visibility_mobile(browser=browser, visible_deck=Decks.URLS)
+    assert_panel_visibility_mobile(browser=browser, visible_deck=Decks.URLS)
